+++ src/component/top-menu/top-menu.component.jsx
@@ -1,31 +1,35 @@
-import React from 'react'
-import { StyledTopMenu, StyledLink, StyledButton, StyledLogoWrapper } from './top-menu.style'
+import React from 'react';
+import { StyledTopMenu, StyledLink, StyledButton, StyledLogoWrapper } from './top-menu.style';
 import { WebInfoState } from '../web-info/web-info.context';
 import { auth, GitHubProvider, db } from '../data/firebase';
 import Avatar from '../avatar';
 import Logo from '../_dumb/logo/logo.component';
 import HeaderTitle from '../_dumb/header-title';
+import { navigate } from '@reach/router';
 
 const TopMenu = () => {
-  const { toggleChat, updateToggleChat, user, updateUser } = WebInfoState()
+  const { toggleChat, updateToggleChat, user, updateUser } = WebInfoState();
 
   const handleToggleChat = () => {
     updateToggleChat({ type: 'TOGGLE_CHAT' });
-  }
+  };
 
   const handleLogInAndOut = () => {
     if (user) {
       auth.signOut();
+      navigate('/');
       updateUser({
         type: 'USER_AUTHENTICATE',
         user: null
-      })
-    }
-    else {
+      });
+    } else {
       auth
         .signInWithPopup(GitHubProvider)
         .then(({ user: { uid, displayName, photoURL, email } }) => {
-          db.collection('user').doc(uid).set({ displayName, photoURL, email }, { merge: true })
+          db.collection('user')
+            .doc(uid)
+            .set({ displayName, photoURL, email }, { merge: true });
+          navigate('/dashboard');
           updateUser({
             type: 'USER_AUTHENTICATE',
             user: {
@@ -33,19 +37,18 @@ const TopMenu = () => {
               displayName,
               photoURL
             }
-          })
+          });
         })
         .catch(error => {
           // console.error(error)
-        })
+        });
     }
-  }
+  };
 
-  const getLogInOutLabel = () => user ? 'Ba Bye' : 'Let me in!'
+  const getLogInOutLabel = () => (user ? 'Ba Bye' : 'Let me in!');
 
   return (
     <StyledTopMenu>
-
       <StyledLink to="/">
         <StyledLogoWrapper>
           <Logo />
@@ -55,16 +58,14 @@ const TopMenu = () => {
           />
         </StyledLogoWrapper>
       </StyledLink>
-      <StyledLink to="/dashboard">Dashboard</StyledLink>
+      {user && <StyledLink to="/dashboard">Dashboard</StyledLink>}
       <Avatar user={user} />
-      <button onClick={handleLogInAndOut}>
-        {getLogInOutLabel()}
-      </button>
-      <StyledButton onClick={handleToggleChat}>{
-        toggleChat ? 'Invizi Chat' : 'Gimme chat now!'
-      }</StyledButton>
+      <button onClick={handleLogInAndOut}>{getLogInOutLabel()}</button>
+      <StyledButton onClick={handleToggleChat}>
+        {toggleChat ? 'Invizi Chat' : 'Gimme chat now!'}
+      </StyledButton>
     </StyledTopMenu>
-  )
-}
+  );
+};
 
-export default TopMenu
+export default TopMenu;